% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roasted.R
\docType{package}
\name{roasted-package}
\alias{roasted}
\alias{roasted-package}
\title{Java Classes In R}
\description{
This package includes classes like those commonly used by Java
programmers. While the R implementations of these classes may not provide
runtime improvements, the package aims to diversify the toolset with which R
programmers can tackle and think about problems.
}

