% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{string}
\alias{is.string}
\alias{string}
\title{Strings}
\usage{
string(value)

is.string(x)
}
\arguments{
\item{value}{The initial value of the string.}

\item{x}{An object to be tested.}

\item{offset}{The initial offset.}

\item{count}{The length.}
}
\description{
Initialize a newly created string object so that it represents an empty
character sequence.
}

