% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{String}
\alias{String}
\title{String}
\usage{
String(value = NULL, offset = 1, count = length(value))
}
\arguments{
\item{value}{The initial value of the string.}

\item{offset}{The initial offset.}

\item{count}{The length.}
}
\description{
Initialize a newly created String object so that it represents an empty
character sequence.
}

